refactor: dto, repository 분리
- #189
