+++ src/ext/ext.module.ts
@@ -3,8 +3,8 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 import { AuthModule } from 'src/auth/auth.module';
 import { DeviceInfo } from 'src/entities/device-info.entity';
 import { TagLog } from 'src/entities/tag-log.entity';
-import { TagLogRepository } from 'src/tag-log-v1/repository/mysql/tag-log.repository';
 import { TagLogModule } from 'src/tag-log-v1/tag-log.module';
+import { TagLogRepository } from 'src/tag-log/repository/mysql/tag-log.repository';
 import { UserModule } from 'src/user/user.module';
 import { ExtService } from './ext.service';
 import { DeviceInfoRepository } from './repository/mysql/device-info.repository';
