--- src/tag-log-v2/dto/tag-log.dto.ts
@@ -1,6 +1,0 @@
-export class TagLogDto {
-  idx: number;
-  tag_at: Date;
-  card_id: string;
-  device_id: number;
-}
