+++ src/tag-log-v2/tag-log-admin.controller.ts
@@ -17,18 +17,18 @@ import {
   ApiResponse,
   ApiTags,
 } from '@nestjs/swagger';
-import { User } from 'src/auth/user.decorator';
 import { UserSessionDto } from 'src/auth/dto/user.session.dto';
-import { UserAccumulationDayType } from './dto/admin/user-accumulation-day.type';
-import { UserAccumulationMonthType } from './dto/admin/user-accumulation-month.type';
-import { TagLogAdminService } from './tag-log-v2-admin.service';
 import { AdminAuthGuard } from 'src/auth/guard/admin-auth.guard';
+import { User } from 'src/auth/user.decorator';
+import { UserAccumulationDayType } from 'src/tag-log/dto/admin/user-accumulation-day.type';
+import { UserAccumulationMonthType } from 'src/tag-log/dto/admin/user-accumulation-month.type';
+import { TagLogAdminService } from './tag-log-v2-admin.service';
 
-@ApiTags('체류 시간 산출 V2 (관리자 전용 API)')
+@ApiTags('체류 시간 산출 (관리자 전용 API)')
 @ApiBearerAuth()
 @UseGuards(AdminAuthGuard)
 @Controller({
-  version: '2',
+  version: '1',
   path: 'tag-log/admin',
 })
 export class TagLogAdminController {
@@ -148,7 +148,6 @@ export class TagLogAdminController {
       year,
       month,
     );
-    console.log(result);
     return result;
   }
