+++ src/tag-log-v2/tag-log-v2.module.ts
@@ -5,15 +5,14 @@ import { DeviceInfo } from 'src/entities/device-info.entity';
 import { PairInfo } from 'src/entities/pair-info.entity';
 import { TagLog } from 'src/entities/tag-log.entity';
 import { StatisticsModule } from 'src/statistics/statictics.module';
+import { PairInfoRepository } from 'src/tag-log/repository/mysql/pair-info.repository';
 import { UserModule } from 'src/user/user.module';
 import { UtilsModule } from 'src/utils/utils.module';
 import { Cabi42Controller } from '../ext/cabi42.controller';
-import { DeviceInfoRepository } from './repository/mysql/device-info.repository';
-import { PairInfoRepository } from './repository/mysql/pair-info.repository';
-import { TagLogRepository } from './repository/mysql/tag-log.repository';
-import { TagLogAdminController } from './tag-log-v2-admin.controller';
+import { DeviceInfoRepository } from '../tag-log/repository/mysql/device-info.repository';
+import { TagLogRepository } from '../tag-log/repository/mysql/tag-log.repository';
 import { TagLogAdminService } from './tag-log-v2-admin.service';
-
+import { TagLogAdminController } from './tag-log-admin.controller';
 import { TagLogController } from './tag-log-v2.controller';
 import { TagLogService } from './tag-log-v2.service';
