+++ src/tag-log-v2/tag-log-v2.service.spec.ts
@@ -1,9 +1,9 @@
 import { Test, TestingModule } from '@nestjs/testing';
-import { TagLogService } from './tag-log-v2.service';
-import { PairInfoDto } from './dto/pair-info.dto';
-import { TagLogDto } from './dto/tag-log.dto';
-import { DateCalculator } from 'src/utils/date-calculator.component';
+import { PairInfoDto } from 'src/tag-log/dto/pair-info.dto';
+import { TagLogDto } from 'src/tag-log/dto/tag-log.dto';
 import { UserService } from 'src/user/user.service';
+import { DateCalculator } from 'src/utils/date-calculator.component';
+import { TagLogService } from './tag-log-v2.service';
 
 const mocks = [
   {
