+++ src/tag-log/repository/interface/tag-log-repository.interface.ts
@@ -1,4 +1,4 @@
-import { TagLogDto } from 'src/tag-log-v1/dto/tag-log.dto';
+import { TagLogDto } from 'src/tag-log/dto/tag-log.dto';
 import { CardDto } from 'src/user/dto/card.dto';
 
 export interface ITagLogRepository {
