+++ src/tag-log/repository/mysql/device-info.repository.ts
@@ -1,8 +1,8 @@
-import { DeviceInfoDto } from 'src/tag-log-v2/dto/device-info.dto';
-import { IDeviceInfoRepository } from '../interface/device-info-repository.interface';
 import { InjectRepository } from '@nestjs/typeorm';
-import { Repository } from 'typeorm';
 import { DeviceInfo } from 'src/entities/device-info.entity';
+import { DeviceInfoDto } from 'src/tag-log/dto/device-info.dto';
+import { Repository } from 'typeorm';
+import { IDeviceInfoRepository } from '../interface/device-info-repository.interface';
 
 export class DeviceInfoRepository implements IDeviceInfoRepository {
   constructor(
