+++ src/tag-log/tag-log.service.ts
@@ -1,0 +1,7 @@
+// import { Inject, Injectable, Logger } from '@nestjs/common';
+// import { UserService } from 'src/user/user.service';
+
+// @Injectable()
+// export class TagLogService {
+//   private logger = new Logger(TagLogService.name);
+// }
